Trait matrs::traits::CompliantNumerical

pub trait CompliantNumerical: Sized + Default + Clone + Copy + PartialEq + PartialOrd + Add<Output = Self> + Sub<Output = Self> + Mul<Output = Self> + Div<Output = Self> {
    fn into_f64(self) -> f64;

    fn into_f32(self) -> f32;

    fn into_i8(self) -> i8;

    fn into_i16(self) -> i16;

    fn into_i32(self) -> i32;

    fn into_i64(self) -> i64;

    fn into_u8(self) -> u8;

    fn into_u16(self) -> u16;

    fn into_u32(self) -> u32;

    fn into_u64(self) -> u64;
}

Defines a compliant numerical trait It is used as a generic to only allow numbers that can be casted to/ from floats and integers This trait is used to ensure that the matrix struct can only be used with numbers

Required methods

fn into_f64(self) -> f64

fn into_f32(self) -> f32

fn into_i8(self) -> i8

fn into_i16(self) -> i16

fn into_i32(self) -> i32

fn into_i64(self) -> i64

fn into_u8(self) -> u8

fn into_u16(self) -> u16

fn into_u32(self) -> u32

fn into_u64(self) -> u64