Struct matrs::vec::Vec

pub struct Vec<T: CompliantNumerical, const COUNT: usize> { /* private fields */ }

Type definition for a vector It is a fixed size vector It is used to implement linear algebra for neural nets and similar

Benefits

It is a fixed size vector

It gaurantees that the vector to vector operations are of correct size

Example

use matrs::vec::Vec;
let v1 = Vec::<f32, 3>::new();
let v2 = Vec::<f32, 3>::new();
let v3 = v1 + v2;

Future work

Implementations

impl<T: CompliantNumerical, const COUNT: usize> Vec<T, COUNT>

pub fn new() -> Vec<T, COUNT>

Creates a new vector It is used to create a new vector with a user defined amount of elements

Example

use matrs::vec::Vec;
let vec = Vec::< f64, 3 >::new();

Panics

Never panics

Safety

It is safe to create a new vector

pub fn new_from_data(data: [T; COUNT]) -> Vec<T, COUNT>

Creates a new vector It is used to create a new vector with a user defined amount of elements

Example

use matrs::vec::Vec;
let data:[f32;3] = [1.0, 2.0, 3.0];
let vec = Vec::new_from_data(data);

Panics

Never panics

Safety

It is safe to create a new vector

pub fn get(&self, index: usize) -> &T

Gets the element at the specified index

Example

use matrs::vec::Vec;
let vec = Vec::< f64, 3 >::new();
let element = vec.get(1);

Panics

On index out of bounds

Safety

It is safe if the index is within the bounds of the vector

pub fn get_mut(&mut self, index: usize) -> &mut T

Sets the element at the specified index

Example

use matrs::vec::Vec;
let mut vec = Vec::< f64, 3 >::new();
vec.set(1, 1.0);
let mut element = vec.get(1);

Panics

On index out of bounds

Safety

It is safe if the index is within the bounds of the vector

pub fn set(&mut self, index: usize, value: T)

Sets the element at the specified index

Example

use matrs::vec::Vec;
let mut vec = Vec::< f64, 3 >::new();
vec.set(1, 1.0);

Panics

On index out of bounds

Safety

It is safe if the index is within the bounds of the vector

pub fn len(&self) -> usize

Gets the length of the vector

Example

use matrs::vec::Vec;
let vec = Vec::< f64, 3 >::new();
let length = vec.len();

Panics

Never panics

Safety

It is safe to get the length of the vector

Note

The length of the vector is always the same as the amount of elements in the vector

pub fn iter(&self) -> Iter<'_, T>

Convert the vector to a slice

Example

use matrs::vec::Vec;
let mut vec = Vec::< f64, 3 >::new();
let slice = vec.iter();

Panics

Never panics

Safety

It is safe to convert the vector to a slice

pub fn iter_mut(&mut self) -> IterMut<'_, T>

Convert the vector to a slice

Example

use matrs::vec::Vec;
let mut vec = Vec::< f64, 3 >::new();
let slice = vec.iter_mut();

Panics

Never panics

Safety

It is safe to convert the vector to a slice

pub fn for_each(&self, f: T) where
T: Fn(T) -> T,

Passes each element of the vector to the function